removing circular dependencies from imports
