+++ expressyeaself/build_promoter.py
@@ -5,6 +5,7 @@ For example: ATGCATGC inserted into AAAANNNNNNNNTTTT would give
 AAAAATGCATGCTTTT.
 """
 import expressyeaself.organize_data as organize
+from expressyeaself.utilities import check_valid_line as check_valid_line
 from expressyeaself.utilities import get_time_stamp as get_time_stamp
 from expressyeaself.utilities import smart_open as smart_open
 import os
@@ -111,7 +112,7 @@ def remove_flanks_from_all_seqs(input_seqs, scaffold_type='pTpA'):
     outfile = smart_open(absolute_path, 'w')
     # Remove flanks and write data to output file.
     for line in infile:
-        line = organize.check_valid_line(line)
+        line = check_valid_line(line)
         if line == 'skip_line':
             continue
         seq, exp_level = organize.separate_seq_and_el_data(line)
@@ -207,7 +208,7 @@ def insert_all_seq_into_one_scaffold(input_seqs, scaffold_type='pTpA'):
     scaffold = scaff_file.readline().replace('\n', '')
     # Insert sequences into scaffold and write data to output file
     for line in infile:
-        line = organize.check_valid_line(line)
+        line = check_valid_line(line)
         if line == 'skip_line':
             continue
         seq, exp_level = organize.separate_seq_and_el_data(line)
@@ -266,7 +267,7 @@ def pad_sequences(input_seqs, pad_front=False, extra_padding=0):
     pad_length = max_length + extra_padding
     with smart_open(input_seqs) as f:
         for line in f:
-            line = organize.check_valid_line(line)
+            line = check_valid_line(line)
             if line == 'skip_line':
                 continue
             seq, exp_level = organize.separate_seq_and_el_data(line)
