+++ expressyeaself/organize_data.py
@@ -2,6 +2,7 @@
 This script contains functions to organize and split up data based
 on several experimental parameters.
 """
+from expressyeaself.utilities import check_valid_line as check_valid_line
 from expressyeaself.utilities import get_seq_count as get_seq_count
 from expressyeaself.utilities import get_time_stamp as get_time_stamp
 from expressyeaself.utilities import smart_open as smart_open
@@ -33,39 +34,6 @@ def separate_seq_and_el_data(line):
     exp_level = float(data[1])
 
     return seq, exp_level
-
-
-def check_valid_line(line):
-    """
-    Takes an line from an input file containing sequence and
-    expression level data and returns instructions on what to
-    do based on its classification. For example, if the line is
-    a comment or is empty, the function will return 'skip_line'.
-    If the line is encoded into bytes, it will return the
-    decoded line. Not satisfying these conditionals will mean the
-    line is valid, and so will be returned as it was inputted.
-
-    Args:
-    -----
-        line (str or bytes) -- a line from an input file to be
-        checked for validity
-
-    Returns:
-    -----
-        line (str) - if the input line was valid, the decoded line
-        is returned. Otherwise, the string 'skip_line' will be
-        returned.
-    """
-    if isinstance(line, bytes):  # decodes line if encoded
-        line = line.decode()
-    if line is None or line == "" or line[0] == "#":
-        return 'skip_line'
-    try:
-        seq, exp_level = separate_seq_and_el_data(line)
-    except IndexError:
-        line = 'skip_line'
-
-    return line
 
 
 def get_max_min_mode_length_of_seqs(input_seqs):
